Added validation for API endpoints
Added validation for API endpoints where data could be entered in the request body of an HTTP request, in "index.js file".

In "models.js", hashing was added to the password field of the mongoose schema created for "users" collection, for encrypting the user's input data.

An additional callback function was added to "passport.js" file, which is combined with  the HTTP post request  in "index.js" file. Creating further logic on the hashed passwords sent by a user.

The middleware functions used for the the encrypting and validation of the HTTP requests, were imported from node modules, seen in "package.json" file.