hosted the mongoDB database and the API
The database, previously stored via mongo shell, is now connected online to MongoDB Atlas.
The API, from "index.js" file, is also hosted online, via Heroku.
In "index.js" file, an environment variable (created on Heroku) is
referencing the hosted database, on line 12.

Line 229-231, the listening port was hardcoded to 8080.
Now the app is hosted online, the API needs to look for a pre-defined port in the environment variables, as shown from line 230 in "index.js" file.