Exposed the environment variable
Since this is a portfolio project, environment variable being exposed in the `README.md` profile, and used in the `index.js` makes it considerably easier for contributors to help maintain and update this codebase.

More information as to why this architectual decision has been made, is found on the `README.md`.