feat: add s3 file uploads for localstack integration
packages installed in `package.json` file are for configuring the file upload on the s3 bucket in localstack,.

Installed `dotenv` to configure `.env` file in `index.js`. This `.env` file is added to `gitignore`, and is holding the connection variable of the mongoDB Atlas database, for localhost to gain access to the Atlas database.

S3 bucket endpoint were created in the `index.js` file:
* a GET request to view the list of objects in the s3 bucket on localstack

* A POST request to upload a file to the s3 bucket on localstack